Update Visual Studio path and Windows version in workflow
The commit changes the 'MSVC_PATH' to point to the 2019 version of Microsoft Visual Studio instead of the 2022 version. Correspondingly, the 'os' parameter in the workflow configuration has been adjusted to use 'windows-2019' instead of 'windows-latest'. This should ensure the compatibility of the build process with the 2019 version of the tools.